Recommend `-S` on hashbang
The previous hashbang did not work on Linux coreutils `env` because `env` passes the arguments as single string to `node`. By adding `-S` we can make it split the arguments and this works on both Linux with coreutils >= 8.30 and Mac.

Also see https://github.com/privatenumber/tsx/pull/523 and https://unix.stackexchange.com/a/477651.