fix(register/esm): pass file urls as paths
compile() expects a filename, not a url: it passes the
value directly to the bindings module, which immediately
parses it as a `Path` unconditionally.

This causes input file source maps to fail to load, as it
attempts to find the URL as if it were a file on disk.
(see https://github.com/swc-project/swc/pull/9422)

This is a targeted fix which hopefully minimizes the chance of things breaking.