feat: cache readDefaultTsConfig results
Previously, readDefaultTsConfig would read the tsconfig.json file every
time it was called, which would be at least once for every file being
compiled. By caching the results so that the tsconfig.json file only
needs to be read once, build times can be dramatically reduced, bringing
them in-line with build times when using a hardcoded config.

As an example, on a personal project, this shaves around 25 seconds off
of a 40 second build, meaning it speeds up the build by more than two
times.